Wait for image to finish loading before drawing into canvas
In imgExport(), we export the back image and create an Image() with src as the back image. However, in some situation (especially on Firefox and iOS), the image hasn't finished loading when we draw it onto the canvas. To fix this, we only draw onto the canvas once the image has loaded using the onload callback. As the function is now async, we add a callback function which is called with the Data URL of the finished canvas.